requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.16.0,<3.0.0
google-auth-oauthlib>=1.0.0
safetensors>=0.3.0
Pillow>=9.0.0
numpy>=1.20.0
torch>=2.0.0
